chore: mv functions to their respective files
