+++ app/src/db/users.ts
@@ -8,11 +8,20 @@ import {
   UserInteraction,
 } from "@prisma/client"
 import { CONTRIBUTOR_ELIGIBLE_PROJECTS } from "eas-indexer/src/constants"
+import { TAG_BY_ENTITY } from "eas-indexer/src/constants"
 import { AggregatedType } from "eas-indexer/src/types"
 
 import { UserAddressSource } from "@/lib/types"
+import { mergeResultsByEmail } from "@/lib/utils/tags"
 
 import { prisma } from "./client"
+
+export type Entity = keyof AggregatedType
+export type EntityObject = {
+  address: string
+  email: string
+}
+export type EntityRecords = Record<Entity, EntityObject[]>
 
 export async function getUserById(userId: string) {
   return prisma.user.findUnique({
@@ -568,5 +577,121 @@ export async function getAllCommunityContributors(addresses: string[]) {
         },
       },
     },
+  })
+}
+
+export async function addTag(addresses: EntityObject[], tag: Entity) {
+  if (addresses.length === 0) {
+    return
+  }
+
+  let updatedUserTag = TAG_BY_ENTITY[tag] || ""
+  if (!updatedUserTag) {
+    console.error("Invalid tag")
+    return
+  }
+
+  const usersToUpdate = await prisma.userEmail.findMany({
+    where: {
+      email: {
+        in: addresses.map((a) => a.email),
+      },
+      NOT: {
+        tags: {
+          has: updatedUserTag,
+        },
+      },
+    },
+    select: {
+      email: true,
+      tags: true,
+    },
+  })
+
+  if (usersToUpdate.length === 0) {
+    return await prisma.userEmail.findMany({
+      where: {
+        email: {
+          in: addresses.map((a) => a.email),
+        },
+      },
+      select: {
+        email: true,
+        tags: true,
+      },
+    })
+  }
+
+  await prisma.userEmail.updateMany({
+    where: {
+      email: {
+        in: usersToUpdate.map((u) => u.email),
+      },
+    },
+    data: {
+      tags: {
+        push: updatedUserTag,
+      },
+    },
+  })
+
+  const updatedUsersTags = await prisma.userEmail.findMany({
+    where: {
+      email: {
+        in: usersToUpdate.map((u) => u.email),
+      },
+    },
+    select: {
+      email: true,
+      tags: true,
+    },
+  })
+
+  return mergeResultsByEmail([updatedUsersTags])
+}
+
+export async function removeTags(addresses: EntityRecords, tags: Entity[]) {
+  if (
+    !tags.length ||
+    Object.values(addresses).every((list) => list.length === 0)
+  )
+    return
+
+  const tagsToRemove = tags.map((tag) => TAG_BY_ENTITY[tag]).filter(Boolean)
+  if (!tagsToRemove.length) return
+
+  const emailsToUpdate = Array.from(
+    new Set(
+      Object.values(addresses)
+        .flat()
+        .map((a) => a.email),
+    ),
+  )
+  const usersToUpdate = await prisma.userEmail.findMany({
+    where: {
+      email: { in: emailsToUpdate },
+      tags: { hasSome: tagsToRemove },
+    },
+    select: {
+      id: true,
+      email: true,
+      tags: true,
+    },
+  })
+
+  await prisma.$transaction(
+    usersToUpdate.map((user) =>
+      prisma.userEmail.update({
+        where: { id: user.id },
+        data: {
+          tags: { set: user.tags.filter((t) => !tagsToRemove.includes(t)) },
+        },
+      }),
+    ),
+  )
+
+  return await prisma.userEmail.findMany({
+    where: { email: { in: emailsToUpdate } },
+    select: { email: true, tags: true },
   })
 }
